requirements-1.txt
aiogram==3.13.1
openai==1.57.0
python-dotenv==1.0.1
